refactor error handling

+++ controllers/comments.js
@@ -3,16 +3,16 @@ mongoose.Promise = Promise;
 const { Comments } = require('../models/models');
 const { ObjectId } = require('mongoose').Types;
 
-const getAllComments = (req, res) => {
+const getAllComments = (req, res, next) => {
 	Comments.find()
 		.then(comments => res.status(200).json(comments))
 		.catch(next);
 };
 
 const getCommentById = (req, res, next) => {
-	const { comment_id } = req.params
+	const { comment_id } = req.params;
 	if (!ObjectId.isValid(comment_id)) {
-		const err = new Error(`Comment id is not valid!`);
+		const err = new Error('Comment id is not valid!');
 		err.status = 400;
 		next(err);
 	} else {
@@ -20,53 +20,50 @@ const getCommentById = (req, res, next) => {
 			.then(comment => res.status(200).json(comment))
 			.catch(next);
 	}
-}
+};
 
-const getCommentsByArticle = (req, res) => {
-
+const getCommentsByArticle = (req, res, next) => {
 	Comments.find({ belongs_to: req.params.article_id })
 		.then(comments => {
 			if (!comments) {
-					const err = new Error("Article doesn't exist");
-					err.status = 400;
-					next(err);
-				} else {
-					res.status(200).json(comments)
-						.catch(next);
-				}
-		})
-};
-
-
-const addNewComment = (req, res) => {
-	const { belongs_to, body } = req.body;
-	const comment = new Comments({
-		body,
-		belongs_to: req.params.article_id
-	});
-	comment.save()
-		.then(comment => res.status(201).json(comment))
-		.catch(next);
-};
-
-const voteComment = (req, res) => {
-	let vote = 0;
-	if (req.query.vote === 'up') vote++;
-	else if (req.query.vote === 'down') vote--;
-	Comments.findByIdAndUpdate(req.params.comment_id, { $inc: { votes: vote } }).lean()
-		.then(comment => {
-			comment.votes += vote;
-			res.json(comment);
+				const err = new Error('Article doesn\'t exist');
+				err.status = 400;
+				next(err);
+			} else res.status(200).json(comments);
 		}).catch(next);
 };
 
+const voteComment = (req, res, next) => {
+	const { vote } = req.query;
+	if (vote === 'up' || vote === 'down') {
+		let addVote = 0;
+		if (vote === 'up') addVote++;
+		else if (vote === 'down') addVote--;
+		Comments.findByIdAndUpdate(req.params.comment_id, { $inc: { votes: addVote } }).lean()
+			.then(comment => {
+				comment.votes += addVote;
+				res.json(comment);
+			}).catch(next);
+	}
+	else {
+		const err = new Error('Please provide a query in the format vote=up or vote=down');
+		err.status = 400;
+		next(err);
+	}
+};
 
-const deleteComment = (req, res) => {
+const deleteComment = (req, res, next) => {
 	const { comment_id } = req.params;
-	Comments.findByIdAndRemove(comment_id)
-		.then(message => res.json({ message: `comment ${comment_id} deleted` }))
-		.catch(next);
+	if (!ObjectId.isValid(comment_id)) {
+		const err = new Error('Comment id is not valid!');
+		err.status = 400;
+		next(err);
+	} else {
+		Comments.findByIdAndRemove(comment_id)
+			.then(() => res.json({ message: `comment ${comment_id} deleted` }))
+			.catch(next);
+	}
 };
 
 
-module.exports = { getAllComments, getCommentsByArticle, getCommentById, addNewComment, voteComment, deleteComment };
+module.exports = { getAllComments, getCommentsByArticle, getCommentById, voteComment, deleteComment };
